+++ examples/data-sources/pingfederate_idp_adapter/data-source.tf
@@ -1,0 +1,3 @@
+data "pingfederate_idp_adapter" "idpAdapterExample" {
+  adapter_id = "example"
+}